fastapi_base/pyproject.toml
[project]
name = "fastapi-base"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.16.1",
    "black>=25.1.0",
    "fastapi[standard]>=0.115.12",
    "pydantic-settings>=2.9.1",
    "sentry-sdk[fastapi]>=2.30.0",
    "sqlalchemy>=2.0.41",
]
